Move AddEntryToTree to shared source
