+++ src/Microsoft.AspNetCore.Dispatcher/Tree/TreeMatcher.cs
@@ -175,7 +175,7 @@ namespace Microsoft.AspNetCore.Dispatcher
 
                 var tree = trees[entry.Order];
 
-                AddEntryToTree(tree, entry);
+                TreeEnumerator.AddEntryToTree(tree, entry);
             }
 
             return new Cache(trees.ToArray());
@@ -227,169 +227,6 @@ namespace Microsoft.AspNetCore.Dispatcher
                 }
             }
             return entry;
-        }
-
-        internal static void AddEntryToTree(UrlMatchingTree tree, InboundRouteEntry entry)
-        {
-            // The url matching tree represents all the routes asociated with a given
-            // order. Each node in the tree represents all the different categories
-            // a segment can have for which there is a defined inbound route entry.
-            // Each node contains a set of Matches that indicate all the routes for which
-            // a URL is a potential match. This list contains the routes with the same
-            // number of segments and the routes with the same number of segments plus an
-            // additional catch all parameter (as it can be empty).
-            // For example, for a set of routes like:
-            // 'Customer/Index/{id}'
-            // '{Controller}/{Action}/{*parameters}'
-            //
-            // The route tree will look like:
-            // Root ->
-            //     Literals: Customer ->
-            //                   Literals: Index ->
-            //                                Parameters: {id}
-            //                                                Matches: 'Customer/Index/{id}'
-            //     Parameters: {Controller} ->
-            //                     Parameters: {Action} ->
-            //                                     Matches: '{Controller}/{Action}/{*parameters}'
-            //                                     CatchAlls: {*parameters}
-            //                                                    Matches: '{Controller}/{Action}/{*parameters}'
-            //
-            // When the tree router tries to match a route, it iterates the list of url matching trees
-            // in ascending order. For each tree it traverses each node starting from the root in the
-            // following order: Literals, constrained parameters, parameters, constrained catch all routes, catch alls.
-            // When it gets to a node of the same length as the route its trying to match, it simply looks at the list of
-            // candidates (which is in precence order) and tries to match the url against it.
-
-            var current = tree.Root;
-            var matcher = new RoutePatternMatcher(entry.RoutePattern, entry.Defaults);
-
-            for (var i = 0; i < entry.RoutePattern.PathSegments.Count; i++)
-            {
-                var segment = entry.RoutePattern.PathSegments[i];
-                if (!segment.IsSimple)
-                {
-                    // Treat complex segments as a constrained parameter
-                    if (current.ConstrainedParameters == null)
-                    {
-                        current.ConstrainedParameters = new UrlMatchingNode(depth: i + 1);
-                    }
-
-                    current = current.ConstrainedParameters;
-                    continue;
-                }
-
-                Debug.Assert(segment.Parts.Count == 1);
-                var part = segment.Parts[0];
-                if (part.IsLiteral)
-                {
-                    var literal = (RoutePatternLiteral)part;
-                    if (!current.Literals.TryGetValue(literal.Content, out var next))
-                    {
-                        next = new UrlMatchingNode(depth: i + 1);
-                        current.Literals.Add(literal.Content, next);
-                    }
-
-                    current = next;
-                    continue;
-                }
-
-                // We accept templates that have intermediate optional values, but we ignore
-                // those values for route matching. For that reason, we need to add the entry
-                // to the list of matches, only if the remaining segments are optional. For example:
-                // /{controller}/{action=Index}/{id} will be equivalent to /{controller}/{action}/{id}
-                // for the purposes of route matching.
-                if (part.IsParameter &&
-                    RemainingSegmentsAreOptional(entry.RoutePattern.PathSegments, i))
-                {
-                    current.Matches.Add(new InboundMatch() { Entry = entry, RoutePatternMatcher = matcher });
-                }
-
-                var parameter = part as RoutePatternParameter;
-                if (parameter != null && parameter.Constraints.Any() && !parameter.IsCatchAll)
-                {
-                    if (current.ConstrainedParameters == null)
-                    {
-                        current.ConstrainedParameters = new UrlMatchingNode(depth: i + 1);
-                    }
-
-                    current = current.ConstrainedParameters;
-                    continue;
-                }
-
-                if (parameter != null && !parameter.IsCatchAll)
-                {
-                    if (current.Parameters == null)
-                    {
-                        current.Parameters = new UrlMatchingNode(depth: i + 1);
-                    }
-
-                    current = current.Parameters;
-                    continue;
-                }
-
-                if (parameter != null && parameter.Constraints.Any() && parameter.IsCatchAll)
-                {
-                    if (current.ConstrainedCatchAlls == null)
-                    {
-                        current.ConstrainedCatchAlls = new UrlMatchingNode(depth: i + 1) { IsCatchAll = true };
-                    }
-
-                    current = current.ConstrainedCatchAlls;
-                    continue;
-                }
-
-                if (parameter != null && parameter.IsCatchAll)
-                {
-                    if (current.CatchAlls == null)
-                    {
-                        current.CatchAlls = new UrlMatchingNode(depth: i + 1) { IsCatchAll = true };
-                    }
-
-                    current = current.CatchAlls;
-                    continue;
-                }
-
-                Debug.Fail("We shouldn't get here.");
-            }
-
-            current.Matches.Add(new InboundMatch() { Entry = entry, RoutePatternMatcher = matcher });
-            current.Matches.Sort((x, y) =>
-            {
-                var result = x.Entry.Precedence.CompareTo(y.Entry.Precedence);
-                return result == 0 ? x.Entry.RoutePattern.RawText.CompareTo(y.Entry.RoutePattern.RawText) : result;
-            });
-        }
-
-        private static bool RemainingSegmentsAreOptional(IReadOnlyList<RoutePatternPathSegment> segments, int currentParameterIndex)
-        {
-            for (var i = currentParameterIndex; i < segments.Count; i++)
-            {
-                if (!segments[i].IsSimple)
-                {
-                    // /{complex}-{segment}
-                    return false;
-                }
-
-                var part = segments[i].Parts[0];
-                if (!part.IsParameter)
-                {
-                    // /literal
-                    return false;
-                }
-
-                var parameter = (RoutePatternParameter)part;
-                var isOptionlCatchAllOrHasDefaultValue = parameter.IsOptional ||
-                    parameter.IsCatchAll ||
-                    parameter.DefaultValue != null;
-
-                if (!isOptionlCatchAllOrHasDefaultValue)
-                {
-                    // /{parameter}
-                    return false;
-                }
-            }
-
-            return true;
         }
 
         private struct Key : IEquatable<Key>
